Don't set the X-Frame-Options header
Sinatra sets the X-Frame-Options header to SAMEORIGIN by default.
This means we can't display this dashboard inside a frame.

I can't think of a reason why we might want to keep this enabled
for this app, as it's a dashboard.

This explicitly disables the X-Frame-Options header as part of
the initial Sinatra configuration.